Moves code to netstandard 2.1
Attempts to use EventStore GRPC client on Blazor apps fail because the application is dependent on netcoreapp3.1 to set the HttpClient "Version" property.  With a delegating handler added to the pipeline where the HttpRequestMessage can be accessed as part of building the final request, we are able to set the version each time.