Trigger rediscovery when failing to send a message on a streaming call
- Duplex streaming and client streaming.

Before this, persistent subscriptions could keep retrying subscribing to a down node without rediscovering.
This could happen if the original subscription failed while sending an ack (we were not triggering a rediscovery on failed send)
and then on retry we send the ReadReq (which would fail, not triggering rediscovery) before trying to read from the server.